requirements.txt
langchain-core
langchain-openai
langchain-community
openai
python-dotenv
pydantic
fastapi
uvicorn
streamlit
pymongo